requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.3
opencv-python>=4.9.0
pillow>=10.0.0
tqdm>=4.66.0
PyYAML>=6.0.1
triton-windows; platform_system == "Windows"
